+++ apps/backend/src/app/api/latest/internal/projects/current/crud.tsx
@@ -1,15 +1,21 @@
 import { createOrUpdateProject } from "@/lib/projects";
 import { getTenancy } from "@/lib/tenancies";
-import { retryTransaction } from "@/prisma-client";
+import { getPrismaClientForTenancy, globalPrismaClient } from "@/prisma-client";
 import { createCrudHandlers } from "@/route-handlers/crud-handler";
 import { projectsCrud } from "@stackframe/stack-shared/dist/interface/crud/projects";
 import { yupObject } from "@stackframe/stack-shared/dist/schema-fields";
-import { throwErr } from "@stackframe/stack-shared/dist/utils/errors";
+import { StatusError, throwErr } from "@stackframe/stack-shared/dist/utils/errors";
 import { createLazyProxy } from "@stackframe/stack-shared/dist/utils/proxies";
 
 export const projectsCrudHandlers = createLazyProxy(() => createCrudHandlers(projectsCrud, {
   paramsSchema: yupObject({}),
   onUpdate: async ({ auth, data }) => {
+    if (
+      data.config?.email_theme &&
+      !Object.keys(auth.tenancy.completeConfig.emails.themeList).includes(data.config.email_theme)
+    ) {
+      throw new StatusError(400, "Invalid email theme");
+    }
     const project = await createOrUpdateProject({
       type: "update",
       projectId: auth.project.id,
@@ -29,45 +35,43 @@ export const projectsCrudHandlers = createLazyProxy(() => createCrudHandlers(pro
     };
   },
   onDelete: async ({ auth }) => {
-    await retryTransaction(async (tx) => {
-      await tx.project.delete({
-        where: {
-          id: auth.project.id
-        }
-      });
+    await globalPrismaClient.project.delete({
+      where: {
+        id: auth.project.id
+      }
+    });
 
-      // delete managed ids from users
-      const users = await tx.projectUser.findMany({
-        where: {
-          mirroredProjectId: 'internal',
-          serverMetadata: {
-            path: ['managedProjectIds'],
-            array_contains: auth.project.id
-          }
+    // delete managed ids from users
+    const users = await getPrismaClientForTenancy(auth.tenancy).projectUser.findMany({
+      where: {
+        mirroredProjectId: 'internal',
+        serverMetadata: {
+          path: ['managedProjectIds'],
+          array_contains: auth.project.id
         }
-      });
+      }
+    });
 
-      for (const user of users) {
-        const updatedManagedProjectIds = (user.serverMetadata as any).managedProjectIds.filter(
+    for (const user of users) {
+      const updatedManagedProjectIds = (user.serverMetadata as any).managedProjectIds.filter(
           (id: any) => id !== auth.project.id
         ) as string[];
 
-        await tx.projectUser.update({
-          where: {
-            mirroredProjectId_mirroredBranchId_projectUserId: {
-              mirroredProjectId: 'internal',
-              mirroredBranchId: user.mirroredBranchId,
-              projectUserId: user.projectUserId
-            }
-          },
-          data: {
-            serverMetadata: {
-              ...user.serverMetadata as any,
-              managedProjectIds: updatedManagedProjectIds,
-            }
+      await getPrismaClientForTenancy(auth.tenancy).projectUser.update({
+        where: {
+          mirroredProjectId_mirroredBranchId_projectUserId: {
+            mirroredProjectId: 'internal',
+            mirroredBranchId: user.mirroredBranchId,
+            projectUserId: user.projectUserId
           }
-        });
-      }
-    });
+        },
+        data: {
+          serverMetadata: {
+            ...user.serverMetadata as any,
+            managedProjectIds: updatedManagedProjectIds,
+          }
+        }
+      });
+    }
   }
 }));