+++ apps/e2e/tests/backend/backend-helpers.ts
@@ -689,7 +689,7 @@ export namespace Auth {
       };
     }
 
-    export async function getAuthorizationCode(options: { innerCallbackUrl?: URL, authorizeResponse?: NiceResponse, forceBranchId?: string } = {}) {
+    export async function getMaybeFailingAuthorizationCode(options: { innerCallbackUrl?: URL, authorizeResponse?: NiceResponse, forceBranchId?: string } = {}) {
       let authorizeResponse, innerCallbackUrl;
       if (options.innerCallbackUrl && options.authorizeResponse) {
         innerCallbackUrl = options.innerCallbackUrl;
@@ -706,6 +706,15 @@ export namespace Auth {
           cookie,
         },
       });
+      return {
+        authorizeResponse,
+        innerCallbackUrl,
+        response,
+      };
+    }
+
+    export async function getAuthorizationCode(options: { innerCallbackUrl?: URL, authorizeResponse?: NiceResponse, forceBranchId?: string } = {}) {
+      const { response } = await Auth.OAuth.getMaybeFailingAuthorizationCode(options);
       expect(response).toMatchObject({
         status: 303,
         headers: expect.any(Headers),
@@ -982,7 +991,7 @@ export namespace InternalApiKey {
         ...body,
       },
       headers: {
-        'x-stack-admin-access-token': adminAccessToken ?? (backendContext.value.projectKeys !== "no-project" && backendContext.value.projectKeys.adminAccessToken || throwErr("Missing adminAccessToken")),
+        'x-stack-admin-access-token': adminAccessToken ?? (backendContext.value.projectKeys !== "no-project" && backendContext.value.projectKeys.adminAccessToken || undefined),
       }
     });
     expect(response.status).equals(200);
@@ -1248,21 +1257,22 @@ export namespace User {
         verification_callback_url: "http://localhost:12345/some-callback-url",
       },
     });
-      expect(createUserResponse).toMatchObject({
-        status: 200,
-        body: {
-          access_token: expect.any(String),
-          refresh_token: expect.any(String),
-          user_id: expect.any(String),
-        },
-        headers: expect.anything(),
-      });
-      return {
-        userId: createUserResponse.body.user_id,
-        mailbox,
-        accessToken: createUserResponse.body.access_token,
-        refreshToken: createUserResponse.body.refresh_token,
-      };
+    expect(createUserResponse).toMatchObject({
+      status: 200,
+      body: {
+        access_token: expect.any(String),
+        refresh_token: expect.any(String),
+        user_id: expect.any(String),
+      },
+      headers: expect.anything(),
+    });
+    return {
+      userId: createUserResponse.body.user_id,
+      mailbox,
+      accessToken: createUserResponse.body.access_token,
+      refreshToken: createUserResponse.body.refresh_token,
+      password,
+    };
   }
 
   export async function createMultiple(count: number) {