+++ apps/dev-launchpad/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/dev-launchpad
 
+## 2.8.23
+
+## 2.8.22
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.21
+
+## 2.8.20
+
 ## 2.8.19
 
 ## 2.8.18