+++ examples/cjs-test/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @stackframe/example-cjs-test
 
+## 2.8.23
+
+### Patch Changes
+
+- @stackframe/stack@2.8.23
+
+## 2.8.22
+
+### Patch Changes
+
+- @stackframe/stack@2.8.22
+
+## 2.8.21
+
+### Patch Changes
+
+- @stackframe/stack@2.8.21
+
+## 2.8.20
+
+### Patch Changes
+
+- @stackframe/stack@2.8.20
+
 ## 2.8.19
 
 ### Patch Changes